+++ app/Controllers/TasksController.js
@@ -1,5 +1,6 @@
 import { ProxyState } from "../AppState.js";
 import { tasksService } from "../Services/TasksServices.js";
+import { Pop } from "../Utils/Pop.js";
 
 export class TasksController {
   constructor() {
@@ -22,11 +23,13 @@ export class TasksController {
     }
   }
 
-  deleteTask(id){
+  async deleteTask(id){
+    if (await Pop.confirm('Delete?', 'This will remove your task from the list.', 'question', 'Delete'))
     tasksService.deleteTask(id)
+    Pop.toast('Task successfully deleted', 'success', 'center', 1000)
   }
 
-  checkboxValue() {
-    tasksService.checkboxValue()
+  checkboxValue(taskId) {
+    tasksService.checkboxValue(taskId)
   }
 }
